+++ testkit/testkit.go
@@ -280,6 +280,13 @@ func (tk *TestKit) MustGetErrMsg(sql string, errStr string) {
 	tk.require.EqualError(err, errStr)
 }
 
+// MustContainErrMsg executes a sql statement and assert its error message containing errStr.
+func (tk *TestKit) MustContainErrMsg(sql string, errStr interface{}) {
+	err := tk.ExecToErr(sql)
+	tk.require.Error(err)
+	tk.require.Contains(err.Error(), errStr)
+}
+
 // MustMatchErrMsg executes a sql statement and assert its error message matching errRx.
 func (tk *TestKit) MustMatchErrMsg(sql string, errRx interface{}) {
 	err := tk.ExecToErr(sql)
